MVC integration
