+++ src/com/jalasoft/search/MainSearch.java
@@ -10,7 +10,9 @@ package com.jalasoft.search;
 
 import com.jalasoft.search.controller.Controller;
 import com.jalasoft.search.model.Search;
-import com.jalasoft.search.view.Window;
+import com.jalasoft.search.view.View;
+
+import javax.swing.*;
 
 /**
  * This Class is the main program
@@ -22,9 +24,13 @@ import com.jalasoft.search.view.Window;
 public class MainSearch {
 
     public static void main(String[] args) {
-        Search seach = new Search();
-        Window view = new Window("Searcher");
-        Controller controller = new Controller(seach, view);
-//        controller.init();
+        Search search = new Search();
+        View view = new View("Searcher");
+        view.initView();
+        view.setLocationRelativeTo(null);
+        view.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        view.setSize(1100, 700);
+        view.setVisible(true);
+        Controller controller = new Controller(search, view);
     }
 }
