--- src/com/jalasoft/search/model/readme
@@ -1,1 +1,0 @@
-This folder contains the methods and classes common to the project
